* made some field names more consistent, and worked out that     unknown1 is profile_path
If we are going to use automatic backend database generation then we
are going to have to move to really consistent field names in all our
IDL. Thats the only way we are going to be able to automate it a
lot. That will take a couple of days to do when we get to it.
(This used to be commit 1732aae562927afc1ef4f30ef7787cf2d43063e2)