Patch by Stefan Metzmacher <metze@metzemix.de>:
here's a small fix that fixes the new quota system on irix.

I need to reanable XFS quotas on irix for the new quota system
(Jerry do you want to wait for this for the release ?)

But the old system works and is the default on irix!
(This used to be commit 5d43e00a49afc4cf523a531ae6db1a3a8b86c650)