(merge from 3.0)
Fix for debian Bug#225328 by LaMont Jones <lamont@debian.org>, where
the failure of our LFS test on ia64 caused the _GNU_SOURCE define not to be
added, causeing strndup() not to be defined...

This was due to strdup() bein implicitly declared, and casting
pointer->int->pointer.

I'll into when we really should define _GNU_SOURCE shortly, but the
fix is correct anyway.

Andrew Bartlett