added the -C option to smbtorture
this option takes a filename containing a list of UNC names. When
supplied to a torture test that uses multiple client connections the
test will distribute the load across the listed UNC names. This allows
you to use NBENCH effectively against a server that has multiple IP
addresses, and spread the load among them.