Add and delete aliases via srv_samr_nt. For that I added a RID allocation call to winbindd. idmap_allocate_rid wants information about whether this will be a user or a group, I did not export this to the winbind interface.
The reason for idmap to get that info is to keep consistent with the
algorithmic convention to alloc only even rids for users and odd rids for
groups. I'm not fully convinced that this really gains us anything. Any real
good arguments?

Volker
(This used to be commit 7f62cf933cad69799204bfdc773e08ff0dde0b20)