Apply Craig Barratt's fixes to allow multiple exlusion files and patterns.
Much of this was applied by hand because of the changes in the code. However,
it builds and smbclient seems to work still. The Amanda folks are testing
and it seems to work there as well.
(This used to be commit 10d05b57c5c66718fb1ca476f7214087ddae29ee)