After briefly talking to jerry, commit this:
Redirect clients to shares based on the the client's IP address. The msdfs
link has the following form:

temp1 -> msdfs:@/data/tmp/redir@\temp

The file /data/tmp/redir has the contents:

192.168.234.128 closest
192.168.234 close
        default

Everything in between the @-signs in the link target is replaced by the
substituion in the file.

Jerry, errors in the filename expansion are now answered with
NT_STATUS_OBJECT_NAME_NOT_FOUND

Volker