r963: Tridge, according to ethereal this is an 'Acct Ctrl' field. My tests also indicate this although I could not find any consistent pattern.
I found this as 'net rpc group list local' in Samba3 sets this to 250 and only
gets a fixed, but incomplete list of groups out of W2k3.

I tried to correlate the results I got from w2k3 with the LDAP contents of the
corresponding entries, but I could not find anything. Ethereal only decodes
the lower byte, but to get all it seems necessary to have 0xffff here.

If you have time, could you might want to spend some of it decoding the bits
for SAMR completeness....

Volker