r1541: Slightly restructure cm_open_connection. This now has the infrastructure to fire connect request to any number of DC's at the same time. This is actually done for our own domain when you explicitly have 'password server' set to more than one DC, and this is the case I'm currently most interested in.
This is not the last word on that topic I think. Sooner or later we will have
to get rid of the name status requests and replace them with the GETDC
mailslot requests (the way used by windows to find the DC's name) and fire
those all at the same time when the #1C response from WINS comes in. Then we
might also start to to the CLDAP stuff.

Volker