r1598: This is an alternative approach to connect to LDAP servers. I had problems understanding the OpenLDAP client libraries, so I tried to implement that protocol myself. The central routines are ldap_encode and ldap_decode that convert a struct ldap_message to a byte stream and vice versa. You are free to deliver that to the LDAP server in any way you want, including non-blocking network calls.
For an example how to use the API see utils/net_ads.c, routine
net_ads_test.c. This certainly does not belong there, but this is HEAD, isn't
it? ;-)

Volker

P.S: Not being able to figure out how to do async requests reliably with
OpenLDAP libs was my incentive to do this....