r2015: Reverting fix for #1474.
Jelmer, we need to find another way to solve this bug. This way,
rpcclient is linked to libxml2, libmysqlclient and libpg (with according
dependencies in samba-client.rpm's) if one just wants to build the more
experimental pdb-modules as well.

Guenther
(This used to be commit 67bffc5034121cc0069007a7b79ce189ab0d10be)